Add a 14u wide Harmonic layout that's more similar to ANSI keyboards.
In particular, the relative position of Q to the home keys is the same and we can place all ten letters in the upper row, plus two punctuation characters above Enter.

I don't like the looks of it as much: not as simple and symmetric as the H13c. And it's a whole column wider despite only having two more keys.

And the worst part is that the key positions between this and the H13c are so different that each mapping has to be customized, because our FlexMapping data model is row-based.

Thus: there is stuff to think about while doing something else for a few days!