introduce better marketing name for XHKB.
Thumby Flex highlights its two biggest strengths: thumb keys and easy keycap switching. It avoids the words "ergo" and "wide" in the name, because the montonic stagger is not really ergonomic and the Ergoplank/Ergoboard both have a wider hand position than Thumby Flex. So even if Thumby Flex is wider than ANSI-wide, it's narrowest of my ❤️ layouts.